+++ GRDB/Core/DatabasePool.swift
@@ -160,6 +160,11 @@ public final class DatabasePool: DatabaseWriter {
     }
 }
 
+#if swift(>=5.5) && canImport(_Concurrency)
+// @unchecked because of databaseSnapshotCount and readerPool
+extension DatabasePool: @unchecked Sendable { }
+#endif
+
 extension DatabasePool {
     
     // MARK: - Memory management
@@ -183,15 +188,20 @@ extension DatabasePool {
     /// as much memory as possible.
     private func setupMemoryManagement() {
         let center = NotificationCenter.default
+        
+        // Use raw notification names because of
+        // FB9801372 (UIApplication.didReceiveMemoryWarningNotification should not be declared @MainActor)
+        // TODO: Reuse UIApplication.didReceiveMemoryWarningNotification when possible.
+        // TODO: Reuse UIApplication.didEnterBackgroundNotification when possible.
         center.addObserver(
             self,
             selector: #selector(DatabasePool.applicationDidReceiveMemoryWarning(_:)),
-            name: UIApplication.didReceiveMemoryWarningNotification,
+            name: NSNotification.Name(rawValue: "UIApplicationDidReceiveMemoryWarningNotification"),
             object: nil)
         center.addObserver(
             self,
             selector: #selector(DatabasePool.applicationDidEnterBackground(_:)),
-            name: UIApplication.didEnterBackgroundNotification,
+            name: NSNotification.Name(rawValue: "UIApplicationDidEnterBackgroundNotification"),
             object: nil)
     }
     
@@ -303,6 +313,7 @@ extension DatabasePool: DatabaseReader {
     
     // MARK: - Reading from Database
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func read<T>(_ value: (Database) throws -> T) throws -> T {
         GRDBPrecondition(currentReader == nil, "Database methods are not reentrant.")
         guard let readerPool = readerPool else {
@@ -403,6 +414,7 @@ extension DatabasePool: DatabaseReader {
             }
     }
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func unsafeRead<T>(_ value: (Database) throws -> T) throws -> T {
         GRDBPrecondition(currentReader == nil, "Database methods are not reentrant.")
         guard let readerPool = readerPool else {
@@ -414,6 +426,41 @@ extension DatabasePool: DatabaseReader {
                 return try value(db)
             }
         }
+    }
+    
+    public func asyncUnsafeRead(_ value: @escaping (Result<Database, Error>) -> Void) {
+        // First async jump in order to grab a reader connection.
+        // Honor configuration dispatching (qos/targetQueue).
+        let label = configuration.identifier(
+            defaultLabel: "GRDB.DatabasePool",
+            purpose: "asyncUnsafeRead")
+        configuration
+            .makeDispatchQueue(label: label)
+            .async {
+                do {
+                    guard let readerPool = self.readerPool else {
+                        throw DatabaseError(resultCode: .SQLITE_MISUSE, message: "Connection is closed")
+                    }
+                    let (reader, releaseReader) = try readerPool.get()
+                    
+                    // Second async jump because sync could deadlock if
+                    // configuration has a serial targetQueue.
+                    reader.async { db in
+                        defer {
+                            releaseReader()
+                        }
+                        do {
+                            // The block isolation comes from the DEFERRED transaction.
+                            try db.clearSchemaCacheIfNeeded()
+                            value(.success(db))
+                        } catch {
+                            value(.failure(error))
+                        }
+                    }
+                } catch {
+                    value(.failure(error))
+                }
+            }
     }
     
     public func unsafeReentrantRead<T>(_ value: (Database) throws -> T) throws -> T {
@@ -628,10 +675,12 @@ extension DatabasePool: DatabaseReader {
     
     // MARK: - Writing in Database
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func writeWithoutTransaction<T>(_ updates: (Database) throws -> T) rethrows -> T {
         try writer.sync(updates)
     }
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func barrierWriteWithoutTransaction<T>(_ updates: (Database) throws -> T) rethrows -> T {
         // TODO: throw instead of crashing when the database is closed
         try readerPool!.barrier {