+++ GRDB/Core/DatabaseReader.swift
@@ -19,7 +19,7 @@ import Dispatch
 /// connection to the database. Should you have to cope with external
 /// connections, protect yourself with transactions, and be ready to setup a
 /// [busy handler](https://www.sqlite.org/c3ref/busy_handler.html).
-public protocol DatabaseReader: AnyObject {
+public protocol DatabaseReader: AnyObject, GRDBSendable {
     
     /// The database configuration
     var configuration: Configuration { get }
@@ -148,6 +148,7 @@ public protocol DatabaseReader: AnyObject {
     /// - parameter value: A function that accesses the database.
     /// - throws: The error thrown by `value`, or any `DatabaseError` that would
     ///   happen while establishing the read access to the database.
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     func read<T>(_ value: (Database) throws -> T) throws -> T
     
     /// Asynchronously executes a read-only function that accepts a
@@ -218,7 +219,36 @@ public protocol DatabaseReader: AnyObject {
     /// - parameter value: A function that accesses the database.
     /// - throws: The error thrown by `value`, or any `DatabaseError` that would
     ///   happen while establishing the read access to the database.
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     func unsafeRead<T>(_ value: (Database) throws -> T) throws -> T
+    
+    /// Asynchronously executes a function that accepts a database connection.
+    ///
+    /// The guarantees of the `asyncRead` method are lifted:
+    ///
+    /// the `value` function is not isolated: eventual concurrent database
+    /// updates are visible from the function:
+    ///
+    ///     reader.asyncUnsafeRead { dbResult in
+    ///         do {
+    ///             let db = try dbResult.get()
+    ///             // Those two values can be different, because some other
+    ///             // database connection or some other thread may modify the
+    ///             // database between the two requests.
+    ///             let count1 = try Player.fetchCount(db)
+    ///             let count2 = try Player.fetchCount(db)
+    ///         } catch {
+    ///             // handle error
+    ///         }
+    ///     }
+    ///
+    /// The `value` function is not prevented from writing (DatabaseQueue, in
+    /// particular, will accept database modifications in `asyncUnsafeRead`).
+    ///
+    /// - parameter value: A function that accesses the database. Its argument
+    ///   is a `Result` that provides the database connection, or the failure
+    ///   that would prevent establishing the read access to the database.
+    func asyncUnsafeRead(_ value: @escaping (Result<Database, Error>) -> Void)
     
     /// Synchronously executes a function that accepts a database
     /// connection, and returns its result.
@@ -298,6 +328,102 @@ extension DatabaseReader {
         }
     }
 }
+
+#if compiler(>=5.5.2) && canImport(_Concurrency)
+extension DatabaseReader {
+    // MARK: - Asynchronous Database Access
+    
+    // TODO: remove @escaping as soon as it is possible
+    /// Asynchronously executes a read-only function that accepts a database
+    /// connection, and returns its result.
+    ///
+    /// [**Experimental**](http://github.com/groue/GRDB.swift#what-are-experimental-features)
+    ///
+    /// For example:
+    ///
+    ///     let count = try await reader.read { db in
+    ///         try Player.fetchCount(db)
+    ///     }
+    ///
+    /// The `value` function runs in an isolated fashion: eventual concurrent
+    /// database updates are not visible from the function:
+    ///
+    ///     try await reader.read { db in
+    ///         // Those two values are guaranteed to be equal, even if the
+    ///         // `player` table is modified, between the two requests, by
+    ///         // some other database connection or some other thread.
+    ///         let count1 = try Player.fetchCount(db)
+    ///         let count2 = try Player.fetchCount(db)
+    ///     }
+    ///
+    ///     try await reader.read { db in
+    ///         // Now this value may be different:
+    ///         let count = try Player.fetchCount(db)
+    ///     }
+    ///
+    /// Attempts to write in the database throw a DatabaseError with
+    /// resultCode `SQLITE_READONLY`.
+    ///
+    /// - parameter value: A function that accesses the database.
+    /// - throws: The error thrown by `value`, or any `DatabaseError` that would
+    ///   happen while establishing the read access to the database.
+    @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
+    public func read<T>(_ value: @Sendable @escaping (Database) throws -> T) async throws -> T {
+        try await withUnsafeThrowingContinuation { continuation in
+            asyncRead { result in
+                do {
+                    try continuation.resume(returning: value(result.get()))
+                } catch {
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+    
+    // TODO: remove @escaping as soon as it is possible
+    /// Asynchronously executes a function that accepts a database connection.
+    ///
+    /// [**Experimental**](http://github.com/groue/GRDB.swift#what-are-experimental-features)
+    ///
+    /// For example:
+    ///
+    ///     let count = try await reader.unsafeRead { db in
+    ///         try Player.fetchCount(db)
+    ///     }
+    ///
+    /// The guarantees of the `read` method are lifted:
+    ///
+    /// the `value` function is not isolated: eventual concurrent database
+    /// updates are visible from the function:
+    ///
+    ///     try await reader.asyncRead { db in
+    ///         // Those two values can be different, because some other
+    ///         // database connection or some other thread may modify the
+    ///         // database between the two requests.
+    ///         let count1 = try Player.fetchCount(db)
+    ///         let count2 = try Player.fetchCount(db)
+    ///     }
+    ///
+    /// The `value` function is not prevented from writing (DatabaseQueue, in
+    /// particular, will accept database modifications in `asyncUnsafeRead`).
+    ///
+    /// - parameter value: A function that accesses the database.
+    /// - throws: The error thrown by `value`, or any `DatabaseError` that would
+    ///   happen while establishing the read access to the database.
+    @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
+    public func unsafeRead<T>(_ value: @Sendable @escaping (Database) throws -> T) async throws -> T {
+        try await withUnsafeThrowingContinuation { continuation in
+            asyncUnsafeRead { result in
+                do {
+                    try continuation.resume(returning: value(result.get()))
+                } catch {
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+}
+#endif
 
 #if canImport(Combine)
 extension DatabaseReader {
@@ -456,6 +582,7 @@ public final class AnyDatabaseReader: DatabaseReader {
     
     // MARK: - Reading from Database
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func read<T>(_ value: (Database) throws -> T) throws -> T {
         try base.read(value)
     }
@@ -469,8 +596,13 @@ public final class AnyDatabaseReader: DatabaseReader {
         base._weakAsyncRead(value)
     }
     
+    @_disfavoredOverload // SR-15150 Async overloading in protocol implementation fails
     public func unsafeRead<T>(_ value: (Database) throws -> T) throws -> T {
         try base.unsafeRead(value)
+    }
+    
+    public func asyncUnsafeRead(_ value: @escaping (Result<Database, Error>) -> Void) {
+        base.asyncUnsafeRead(value)
     }
     
     public func unsafeReentrantRead<T>(_ value: (Database) throws -> T) throws -> T {